Migrate index manager tests
